Web-config value for hardware and voip dash display:
Added a config value in the xaps-web.properties config file to let the
admin choose if the hardware and voip value should be shown in the unit
dashboard and unit history. At present these values are only compatible
with certain devices (those from Pingcom) and only clutters the view for
an actor not using those devices.

Additionally the menu alternatives for these displays have been
disabled.